+++ scripts/addEvents.js
@@ -1,9 +1,9 @@
 document.addEventListener("DOMContentLoaded", () => {
-  document.documentElement.addEventListener("keydown", (e) =>{
-    if(e.key === "Escape" && randomStats.elementOpened != null){
-      abrirSecao()
+  document.documentElement.addEventListener("keydown", (e) => {
+    if (e.key === "Escape" && randomStats.elementOpened != null) {
+      abrirSecao();
     }
-  })
+  });
 
   // clicar na batata
   const imagemBatata = document.getElementById("imagemBatata");
@@ -14,15 +14,37 @@ document.addEventListener("DOMContentLoaded", () => {
   // navegação das seções
   const navImgs = document.querySelectorAll(".navSecoes img");
 
-  if (navImgs[0]) { navImgs[0].addEventListener("click", () => {abrirSecao("divConfiguracoes", "", [configAberta]);});}
-  if (navImgs[1]) { navImgs[1].addEventListener("click", () => {abrirSecao("divEstatisticas", "", [estatisticaAberta, estatisticas]);});}
-  if (navImgs[2]) { navImgs[2].addEventListener("click", () => {abrirSecao("divConquistas", "", [abriuConquista]);});}
+  if (navImgs[0]) {
+    navImgs[0].addEventListener("click", () => {abrirSecao("divConfiguracoes", "", [configAberta]);});
+    navImgs[0].addEventListener("mouseenter", () => {hoverAba(0)});
+    navImgs[0].addEventListener("mouseleave", () => {hoverAba()});
+  }
+  if (navImgs[1]) {
+    navImgs[1].addEventListener("click", () => {abrirSecao("divEstatisticas", "", [estatisticaAberta, estatisticas]);});
+    navImgs[1].addEventListener("mouseenter", () => {hoverAba(1)});
+    navImgs[1].addEventListener("mouseleave", () => {hoverAba()});
+  }
+  if (navImgs[2]) {
+    navImgs[2].addEventListener("click", () => {abrirSecao("divConquistas", "", [abriuConquista]);});
+    navImgs[2].addEventListener("mouseenter", () => {hoverAba(2)});
+    navImgs[2].addEventListener("mouseleave", () => {hoverAba()});
+  }
+
+  const navImgSkin = document.querySelector("img[src='imagens/skins.png']");
+  if(navImgSkin){
+    navImgSkin.addEventListener("click", () => {abrirSecao("divSkins", "", []);});
+    // navImgSkin.addEventListener("mouseenter", () => {hoverAba(1)});
+    // navImgSkin.addEventListener("mouseleave", () => {hoverAba()});
+  }
+
+
+
 
   // botões de quantidade de upgrades
   const botoesUpgrade = document.querySelectorAll(".botaoMudarNumUpgrade");
 
   botoesUpgrade.forEach((botao, index) => {
-    if(botao.parentElement.classList.contains("nomeUpgrade")) return
+    if (botao.parentElement.classList.contains("nomeUpgrade")) return;
     botao.addEventListener("click", () => {
       mudarNumUpgrades(index);
     });
@@ -31,7 +53,7 @@ document.addEventListener("DOMContentLoaded", () => {
   // botões de fechar div
   const botoesFechar = document.querySelectorAll(".botaoFecharDiv");
 
-  botoesFechar.forEach(botao => {
+  botoesFechar.forEach((botao) => {
     botao.addEventListener("click", () => {
       abrirSecao();
     });
@@ -66,7 +88,7 @@ document.addEventListener("DOMContentLoaded", () => {
       randomStats.saveManual++;
     });
   }
-  
+
   // Escolha de arquivo para importar save
   const inputImportarSave = document.getElementById("importar-save");
   if (inputImportarSave) {
@@ -82,7 +104,12 @@ document.addEventListener("DOMContentLoaded", () => {
   // Botão para acionar input de anexar arquivo
   const buttonImportSave = document.getElementById("buttonImportSave");
   if (buttonImportSave) {
-    buttonImportSave.addEventListener("click", () => {inputImportarSave.click()});
+    buttonImportSave.addEventListener("click", () => {
+      inputImportarSave.click();
+    });
   }
 
+  document.getElementById("selectBatatasPorSegundo").addEventListener("change", () => {
+    document.getElementById("estatBatatasPorSegundo").innerHTML = valorSelectCPS();
+  });
 });